Core Functionality
This commit includes the core functionality of XMLMapper:
the Model class and the descriptor classes which map
XML to python values and vice versa.

Additionally, an example of the the various mappings is
included in the form of 'example.py', as well as a tutorial
on how to use the various classes.

Finally, there are a series of helpers which facilitate the
creation of common loaders and dumpers.  The helper `create_element`
was also introduced.  It simply proxies its arguments to
etree.Element.  This allows users to create elements without having
to import lxml.  Additionally, it will allow for the switching of
etree implementations without requiring users to rewrite their programs,
should the need arise.